feat: add google-storage-go buffer reuse patch
This patch adds the hability to specify a user-defined buffer to be
used for file uploads to Google Storage in order to reduce memory
allocations. The user needs to keep track of this buffer somehow,
normally with sync.Buffer, and ensure it's not used by multiple
goroutines.

This patch depends on this one from googleapis:
https://github.com/googleapis/google-api-go-client/pull/632